Move DocumentationTarget to own file
