+++ Sources/App/Controllers/PackageController+PackageResult.swift
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 Dave Verwer, Sven A. Schmidt, and other contributors.
+// Copyright 2020-2022 Dave Verwer, Sven A. Schmidt, and other contributors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -63,142 +63,6 @@ extension PackageController.PackageResult {
             .first()
             .unwrap(or: Abort(.notFound))
         return Self.init(model: model)
-    }
-}
-
-//extension PackageController.PackageResult {
-//    var defaultDocumentationUrl: String? {
-//        guard let repositoryOwner = repository.owner,
-//              let repositoryName = repository.name
-//        else { return nil }
-//
-//        if let spiManifest = defaultBranchVersion.spiManifest,
-//           let externalDocumentationString = spiManifest.externalLinks?.documentation,
-//           let externalDocumentationUrl = URL(string: externalDocumentationString) {
-//            // External documentation links have priority over generated documentation.
-//            return externalDocumentationUrl.absoluteString
-//        } else if let releaseVersion = releaseVersion,
-//                  let releaseVersionDocArchive = releaseVersion.docArchives?.first {
-//            // Ideal case is that we have a stable release documentation.
-//            return DocumentationPageProcessor.relativeDocumentationURL(
-//                owner: repositoryOwner,
-//                repository: repositoryName,
-//                reference: "\(releaseVersion.reference)",
-//                docArchive: releaseVersionDocArchive.name)
-//        } else if let defaultBranchDocArchive = defaultBranchVersion.docArchives?.first {
-//            // Fallback is default branch documentation.
-//            return DocumentationPageProcessor.relativeDocumentationURL(
-//                owner: repositoryOwner,
-//                repository: repositoryName,
-//                reference: "\(defaultBranchVersion.reference)",
-//                docArchive: defaultBranchDocArchive.name)
-//        } else {
-//            // There is no default dodcumentation.
-//            return nil
-//        }
-//    }
-//}
-
-//extension PackageController.PackageResult {
-//    var documentationInfo: (reference: String, archive: String)? {
-//        if let spiManifest = defaultBranchVersion.spiManifest,
-//           spiManifest.externalLinks?.documentation != nil {
-//            // If there's external documentation, we're not handling it ourselves - hence the info is nil.
-//            return nil
-//        }
-//
-//        if let releaseVersion = releaseVersion,
-//           let releaseVersionDocArchive = releaseVersion.docArchives?.first {
-//            return ("\(releaseVersion.reference)", releaseVersionDocArchive.name)
-//        }
-//
-//        if let defaultBranchDocArchive = defaultBranchVersion.docArchives?.first {
-//            return ("\(defaultBranchVersion.reference)", defaultBranchDocArchive.name)
-//        }
-//
-//        return nil
-//    }
-//}
-
-
-
-enum DocumentationTarget: Equatable {
-    case external(url: String)
-    case `internal`(owner: String, repository: String, reference: String, archive: String)
-
-    static func query(on database: Database, owner: String, repository: String) async throws -> Self? {
-        let results = try await Joined3<Version, Package, Repository>
-            .query(on: database,
-                   join: \Version.$package.$id == \Package.$id, method: .inner,
-                   join: \Package.$id == \Repository.$package.$id, method: .inner)
-            .filter(Repository.self, \.$owner, .custom("ilike"), owner)
-            .filter(Repository.self, \.$name, .custom("ilike"), repository)
-            .group(.or) {
-                $0
-                    .filter(\Version.$docArchives != nil)
-                    .filter(\Version.$spiManifest != nil)
-            }
-            .field(Repository.self, \.$ownerName)
-            .field(Version.self, \.$commitDate)
-            .field(Version.self, \.$docArchives)
-            .field(Version.self, \.$latest)
-            .field(Version.self, \.$packageName)
-            .field(Version.self, \.$publishedAt)
-            .field(Version.self, \.$reference)
-            .field(Version.self, \.$spiManifest)
-            .all()
-
-        return results
-            .map(\.model)
-            .documentationTarget(owner: owner, repository: repository)
-    }
-}
-
-
-extension [Version] {
-    var defaultBranchVersion: Version? { filter { $0.latest == .defaultBranch}.first }
-    var releaseVersion: Version? { filter { $0.latest == .release}.first }
-
-    func documentationTarget(owner: String, repository: String) -> DocumentationTarget? {
-        // External documentation links have priority over generated documentation.
-        if let spiManifest = defaultBranchVersion?.spiManifest,
-           let documentation = spiManifest.externalLinks?.documentation {
-            return .external(url: documentation)
-        }
-
-        // Ideal case is that we have a stable release documentation.
-        if let version = releaseVersion,
-           let archive = version.docArchives?.first?.name {
-            return .internal(owner: owner,
-                             repository: repository,
-                             reference: "\(version.reference)",
-                             archive: archive)
-        }
-
-        // Fallback is default branch documentation.
-        if let version = defaultBranchVersion,
-           let archive = version.docArchives?.first?.name {
-            return .internal(owner: owner,
-                             repository: repository,
-                             reference: "\(version.reference)",
-                             archive: archive)
-        }
-
-        // There is no default dodcumentation.
-        return nil
-    }
-
-}
-
-
-extension PackageController.PackageResult {
-    var hasDocumentation: Bool { documentationTarget != nil }
-
-    var documentationTarget: DocumentationTarget? {
-        guard let owner = repository.owner, let repo = repository.name else { return .none }
-        return [defaultBranchVersion.model, releaseVersion?.model, preReleaseVersion?.model]
-            .compactMap { $0 }
-            .documentationTarget(owner: owner, repository: repo)
     }
 }
